packages/pyraytrace/pyproject.toml
[project]
name = "pyraytrace"
version = "0.1.0"
description = "3D ionospheric ray tracing — Python port of OT Report 75-76"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
pyraytrace = "pyraytrace.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
